feat(translations): functions for getting texts and particularly translations. Use the new function in TranslationBox.
